Extract utils and class so we can test at boundaries
