+++ utils/index.js
@@ -1,5 +1,10 @@
 const emitFile = require("./emitFile");
+const createConfig = require("./createConfig");
+
+const DEBUG_STRING = "[eleventy-load]";
 
 module.exports = {
   emitFile,
+  createConfig,
+  DEBUG_STRING,
 };
